Split instantiate functionality into instantiate.js
